requirements.txt

python-dotenv
sqlalchemy
aiogram
utils
python-barcode
asyncpg
protobuf
google-auth-oauthlib
google-api-python-client